refactor: Modal 컴포넌트 이름 변경(Modal → CreditModal)
