+++ src/App.tsx
@@ -2,8 +2,11 @@ import { RouterProvider } from 'react-router-dom';
 
 import { router } from '@/routes/router';
 
+import OverlayTest from './shared/components/overlay/modal/OverlayTest';
+
 function App() {
-  return <RouterProvider router={router} />;
+  // return <RouterProvider router={router} />;
+  return <OverlayTest />;
 }
 
 export default App;
