+++ src/shared/components/overlay/modal/CreditModal.css.ts
