+++ src/shared/components/overlay/modal/OverlayTest.tsx
@@ -1,11 +1,11 @@
 import { overlay } from 'overlay-kit';
 
-import Modal from './Modal';
+import CreditModal from './CreditModal';
 
 const OverlayTest = () => {
   const handleOpen = () => {
     overlay.open(({ unmount }) => (
-      <Modal
+      <CreditModal
         onClose={unmount}
         title={`스타일링 이미지대로 가구를\n추천 받으려면 크레딧이 필요해요`}
       />
